project/Exchange.cpp: moved to .h and .cpp, left only main
